+++ .eslintrc
@@ -14,7 +14,8 @@
 	"globals": {
 		"Config": false, "Monitor": false, "toId": false, "Dex": false, "LoginServer": false,
 		"Users": false, "Punishments": false, "Rooms": false, "Verifier": false, "Chat": false,
-		"Tournaments": false, "Dnsbl": false, "Sockets": false, "TeamValidator": false, "Ladders": false
+		"Tournaments": false, "Dnsbl": false, "Sockets": false, "TeamValidator": false,
+		"TeamValidatorAsync": false, "Ladders": false
 	},
 	"extends": "eslint:recommended",
 	"rules": {
